requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
awscli==1.14.44
bcrypt==3.1.7
blinker==1.4
boto3==1.12.8
botocore==1.15.8
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==6.7
cloud-init==19.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
ec2-hibinit-agent==1.0.0
hibagent==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jmespath==0.9.5
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
oauthlib==2.0.6
olefile==0.45.1
PAM==0.4.2
paramiko==2.7.1
Pillow==5.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
roman==2.0.0
rsa==3.4.2
s3transfer==0.3.3
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.14.0
spur==0.3.21
ssh-import-id==5.7
systemd-python==234
termcolor==1.1.0
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==15.1.0
zope.interface==4.3.2